==> src/20 autoencoders_for_conditional_risk_factors/02_convolutional_denoising_autoencoders.py <==
# Convolutional & Denoising Autoencoders
# The insights from Chapter 17, Convolutional Neural Networks, suggest we incorporate convolutional layers into the
# autoencoder to extract information characteristic of the grid-like structure of image data.
# Source: https://blog.keras.io/building-autoencoders-in-keras.html

from pathlib import Path
import pandas as pd
import numpy as np
from numpy.random import choice

import tensorflow as tf
from tensorflow.keras.layers import Input, Conv2D, MaxPooling2D, UpSampling2D
from tensorflow.keras.models import Model
from tensorflow.keras.callbacks import EarlyStopping, ModelCheckpoint
from tensorflow.keras.datasets import fashion_mnist

import matplotlib.pyplot as plt
import seaborn as sns
import warnings

idx = pd.IndexSlice
np.random.seed(seed=42)
tf.random.set_seed(seed=42)
sns.set_style("whitegrid")
plt.rcParams["figure.dpi"] = 300
plt.rcParams["font.size"] = 14
pd.options.display.float_format = "{:,.2f}".format
warnings.filterwarnings("ignore")

gpu_devices = tf.config.experimental.list_physical_devices("GPU")
if gpu_devices:
    print("Using GPU")
    tf.config.experimental.set_memory_growth(gpu_devices[0], True)
else:
    print("Using CPU")

results_path = Path("../data/ch20", "fashion_mnist")
if not results_path.exists():
    results_path.mkdir(parents=True)


if __name__ == "__main__":
    n_classes = 10  # all examples have 10 classes
    cmap = sns.color_palette("Paired", n_classes)

    ## Fashion MNIST Data
    (X_train, y_train), (X_test, y_test) = fashion_mnist.load_data()
    print(X_train.shape, X_test.shape)

    class_dict = {
        0: "T-shirt/top",
        1: "Trouser",
        2: "Pullover",
        3: "Dress",
        4: "Coat",
        5: "Sandal",
        6: "Shirt",
        7: "Sneaker",
        8: "Bag",
        9: "Ankle boot",
    }
    classes = list(class_dict.keys())

    image_size = 28
    ## Reshape & normalize Fashion MNIST data
    def data_prep_conv(x, size=image_size):
        return x.reshape(-1, size, size, 1).astype("float32") / 255

    X_train_scaled = data_prep_conv(X_train)
    X_test_scaled = data_prep_conv(X_test)
    print(X_train_scaled.shape, X_test_scaled.shape)

    ## Combine training steps into function
    def train_autoencoder(path, model, x_train=X_train_scaled, x_test=X_test_scaled):
        callbacks = [
            EarlyStopping(patience=5, restore_best_weights=True),
            ModelCheckpoint(filepath=path, save_best_only=True, save_weights_only=True),
        ]
        model.fit(x=x_train, y=x_train, epochs=100, validation_split=0.1, callbacks=callbacks)
        model.load_weights(path)
        mse = model.evaluate(x=x_test, y=x_test)
        return model, mse

    ## Convolutional Autoencoder
    # We define a three-layer encoder that uses 2D convolutions with 32, 16, and 8 filters, respectively, ReLU
    # activations, and 'same' padding to maintain the input size. The resulting encoding size at the third layer is
    # 4 x 4 x 8 = 128, higher than for the preceding examples:

    ### 3-dim input
    input_ = Input(shape=(28, 28, 1), name="Input_3D")

    ### Encoding Layers
    x = Conv2D(
        filters=32, kernel_size=(3, 3), activation="relu", padding="same", name="Encoding_Conv_1"
    )(input_)
    x = MaxPooling2D(pool_size=(2, 2), padding="same", name="Encoding_Max_1")(x)
    x = Conv2D(
        filters=16, kernel_size=(3, 3), activation="relu", padding="same", name="Encoding_Conv_2"
    )(x)
    x = MaxPooling2D(pool_size=(2, 2), padding="same", name="Encoding_Max_2")(x)
    x = Conv2D(
        filters=8, kernel_size=(3, 3), activation="relu", padding="same", name="Encoding_Conv_3"
    )(x)
    encoded_conv = MaxPooling2D(pool_size=(2, 2), padding="same", name="Encoding_Max_3")(x)

    # We also define a matching decoder that reverses the number of filters and uses 2D upsampling instead of max
    # pooling to reverse the reduction of the filter sizes. The three-layer autoencoder has 12,785 parameters, a little
    # more than 5% of the capacity of the preceding deep autoencoder.
    x = Conv2D(
        filters=8, kernel_size=(3, 3), activation="relu", padding="same", name="Decoding_Conv_1"
    )(encoded_conv)
    x = UpSampling2D(size=(2, 2), name="Decoding_Upsample_1")(x)
    x = Conv2D(
        filters=16, kernel_size=(3, 3), activation="relu", padding="same", name="Decoding_Conv_2"
    )(x)
    x = UpSampling2D(size=(2, 2), name="Decoding_Upsample_2")(x)
    x = Conv2D(filters=32, kernel_size=(3, 3), activation="relu", name="Decoding_Conv_3")(x)
    x = UpSampling2D(size=(2, 2), name="Decoding_Upsample_3")(x)
    decoded_conv = Conv2D(
        filters=1, kernel_size=(3, 3), activation="sigmoid", padding="same", name="Decoding_Conv_4"
    )(x)

    autoencoder_conv = Model(input_, decoded_conv)
    autoencoder_conv.compile(optimizer="adam", loss="mse")
    autoencoder_conv.summary()

    path = (results_path / "autencoder_conv.32.weights.hdf5").as_posix()
    autoencoder_deep, mse = train_autoencoder(
        path, autoencoder_conv, x_train=X_train_scaled, x_test=X_test_scaled
    )

    # Training stops after 75 epochs and results in a further 9% reduction of the test RMSE, due to a combination of
    # the ability of convolutional filters to learn more efficiently from image data and the larger encoding size.
    print(f"MSE: {mse:.4f} | RMSE {mse**.5:.4f}")

    autoencoder_conv.load_weights(path)
    reconstructed_images = autoencoder_deep.predict(X_test_scaled)
    print(reconstructed_images.shape)

    fig, axes = plt.subplots(ncols=n_classes, nrows=2, figsize=(20, 4))
    for i in range(n_classes):
        axes[0, i].imshow(X_test_scaled[i].reshape(image_size, image_size), cmap="gray")
        axes[0, i].axis("off")
        axes[1, i].imshow(reconstructed_images[i].reshape(image_size, image_size), cmap="gray")
        axes[1, i].axis("off")
    plt.savefig("images/02-01.png")

    ## Denoising Autoencoder
    # The application of an autoencoder to a denoising task only affects the training stage. In this example, we add
    # noise to the Fashion MNIST data from a standard normal distribution while maintaining the pixel values in the
    # range of [0, 1], as follows:
    def add_noise(x, noise_factor=0.3):
        return np.clip(x + noise_factor * np.random.normal(size=x.shape), 0, 1)

    X_train_noisy = add_noise(X_train_scaled)
    X_test_noisy = add_noise(X_test_scaled)

    fig, axes = plt.subplots(nrows=2, ncols=5, figsize=(20, 4))
    axes = axes.flatten()
    for i, ax in enumerate(axes):
        ax.imshow(X_test_noisy[i].reshape(28, 28), cmap="gray")
        ax.axis("off")
    plt.savefig("images/02-02.png")

    x = Conv2D(
        filters=32, kernel_size=(3, 3), activation="relu", padding="same", name="Encoding_Conv_1"
    )(input_)
    x = MaxPooling2D(pool_size=(2, 2), padding="same", name="Encoding_Max_1")(x)
    x = Conv2D(
        filters=32, kernel_size=(3, 3), activation="relu", padding="same", name="Encoding_Conv_2"
    )(x)
    encoded_conv = MaxPooling2D(pool_size=(2, 2), padding="same", name="Encoding_Max_3")(x)

    x = Conv2D(
        filters=32, kernel_size=(3, 3), activation="relu", padding="same", name="Decoding_Conv_1"
    )(encoded_conv)
    x = UpSampling2D(size=(2, 2), name="Decoding_Upsample_1")(x)
    x = Conv2D(
        filters=32, kernel_size=(3, 3), activation="relu", padding="same", name="Decoding_Conv_2"
    )(x)
    x = UpSampling2D(size=(2, 2), name="Decoding_Upsample_2")(x)
    decoded_conv = Conv2D(
        filters=1, kernel_size=(3, 3), activation="sigmoid", padding="same", name="Decoding_Conv_4"
    )(x)

    autoencoder_denoise = Model(input_, decoded_conv)
    autoencoder_denoise.compile(optimizer="adam", loss="mse")

    path = (results_path / "autencoder_denoise.32.weights.hdf5").as_posix()
    callbacks = [
        EarlyStopping(patience=5, restore_best_weights=True),
        ModelCheckpoint(filepath=path, save_best_only=True, save_weights_only=True),
    ]

    # We then proceed to train the convolutional autoencoder on noisy input with the objective to learn how to generate
    # the uncorrupted originals:
    autoencoder_denoise.fit(
        x=X_train_noisy,
        y=X_train_scaled,
        epochs=100,
        batch_size=128,
        shuffle=True,
        validation_split=0.1,
        callbacks=callbacks,
    )

    autoencoder_denoise.load_weights(path)
    mse = autoencoder_denoise.evaluate(x=X_test_noisy, y=X_test_scaled)
    print(f"MSE: {mse:.4f} | RMSE {mse**.5:.4f}")

    ## Visualize Reconstructed Images
    # The following figure shows, from top to bottom, the original images as well as the noisy and denoised versions.
    # It illustrates that the autoencoder is successful in producing compressed encodings from the noisy images that
    # are quite similar to those produced from the original images:
    reconstructed_images = autoencoder_denoise.predict(X_test_noisy)
    print(reconstructed_images.shape)

    fig, axes = plt.subplots(ncols=n_classes, nrows=3, figsize=(20, 6))
    for i in range(n_classes):
        axes[0, i].imshow(X_test[i].reshape(image_size, image_size), cmap="gray")
        axes[0, i].axis("off")
        axes[1, i].imshow(X_test_noisy[i].reshape(image_size, image_size), cmap="gray")
        axes[1, i].axis("off")
        axes[2, i].imshow(reconstructed_images[i].reshape(image_size, image_size), cmap="gray")
        axes[2, i].axis("off")
    fig.suptitle("Originals, Corrupted and Reconstructed Images", fontsize=20)
    fig.tight_layout()
    fig.subplots_adjust(top=0.9)
    fig.savefig("images/02_autoencoder_denoising.png", dpi=300)
